EYQB-1564 split out the help journey into partial classes
